+++ src/app/builders/dialog.ts
@@ -1,0 +1,76 @@
+import { ElementBuilder } from "./element";
+import { is } from "../../core/index";
+
+export class DialogBuilder {
+    #header: Element;
+    #body: Element;
+    #footer: Element;
+    #prefix: string;
+    #switches: string;
+    constructor(prefix: string, switches?: string) {
+        this.#prefix = prefix;
+        this.#header = this.#footer = this.#body = null;
+        this.#switches = switches ?? "";
+    }
+
+    createHeader(title: string, classes: string[], elements?: Element[]) {
+        if (!is(classes)) {
+            classes = []
+        }
+        if (!is(elements)) {
+            elements = []
+        }
+        let headerBuilder = new ElementBuilder('div');
+        headerBuilder.setClasses(`${this.#prefix}-dialog-header`, ...classes)
+        this.#header = headerBuilder.build();
+
+        let titleElement = new ElementBuilder('span').setClasses(this.getPrefixedString("-dialog-title")).build(title)
+
+        this.#header.appendChild(titleElement);
+        this.appendChildrens(this.#header, ...elements);
+
+    }
+
+    createFooter(classes: string[], elements: Element[]) {
+        if (!is(classes)) {
+            classes = []
+        }
+        if (!is(elements)) {
+            elements = []
+        }
+        this.#footer = new ElementBuilder('div').setClasses(this.getPrefixedString("-dialog-footer"), ...classes).build();
+        this.appendChildrens(this.#footer, ...elements);
+    }
+
+    createBody(classes: string[], elements: Element[]) {
+        if (!is(classes)) {
+            classes = []
+        }
+        if (!is(elements)) {
+            elements = []
+        }
+        this.#body = new ElementBuilder('div').setClasses(this.getPrefixedString("-dialog-body"), ...classes).build();
+        this.appendChildrens(this.#body, ...elements);
+    }
+
+    build(id: string): HTMLElement {
+        let dialog = new ElementBuilder('div').setId(id).setClasses(this.getPrefixedString("-dialog")).setAttributes({
+            [this.getPrefixedString('-dialog')]: this.#switches
+        }).build();
+        let container = new ElementBuilder('div').setClasses(this.getPrefixedString("-dialog-container")).build();
+        container.appendChild(this.#header);
+        container.appendChild(this.#body);
+        container.appendChild(this.#footer);
+        dialog.appendChild(container);
+        return dialog;
+    }
+
+    private appendChildrens(parent: Element, ...elements: Element[]) {
+        if (is(elements)) {
+            elements.forEach((element) => parent.appendChild(element));
+        }
+    }
+    private getPrefixedString(str: string) {
+        return this.#prefix + str
+    }
+}